add Disney Princess Glitter Magic Lesson (Japan) as NOT WORKING
