+++ src/mame/mame.lst
@@ -45788,6 +45788,7 @@ duelmast                        //
 epo_ebox                        //
 epo_bowl                        //
 epo_sdb                         //
+tom_dpgm                        //
 ttv_sw                          //
 ttv_lotr                        //
 ttv_mx                          //
